file.c polished. Moved process_range() into file.c and unlinked range.c
